created Constants file/class; started action button code
