+++ Interface/LockOptions.cs
@@ -21,7 +21,7 @@ namespace InvisibleHand
             FIXME: No visual feedback when bringing the mouse to hover over the button.
             TODO: This class could still use some cleanup.
         */
-        public LockOptions(TAPI.ModBase mbase) : base("InvisibleHand:LockOptions")
+        public LockOptions(ModBase mbase) : base("InvisibleHand:LockOptions")
         {
             // this should put the Buttons to the left of the chest inventory
             float posX = 73 - (Main.inventoryBackTexture.Width * Main.inventoryScale);
@@ -29,25 +29,26 @@ namespace InvisibleHand
 
             Buttons = new Dictionary<VAction, IHToggle>();
 
-            // lockDA = new IHToggle("dalock", "Deposit All Locked", "Deposit All Unlocked", null, () => {return IHPlayer.daLocked;}, () =>
-            // Buttons[VAction.DA] = new IHToggle("DA", "da", null, () => IHPlayer.ActionLocked(Main.localPlayer, VAction.DA), () => //onToggle
-            Buttons[VAction.DA] = new IHToggle("Deposit All Locked", "Deposit All Unlocked", mbase.textures["resources/btn_depositall"], () => IHPlayer.ActionLocked(Main.localPlayer, VAction.DA), () =>
-                {
+            Buttons[VAction.DA] = new IHToggle("Deposit All Locked", "Deposit All Unlocked",
+                mbase.textures["resources/btn_depositall"],
+                () => IHPlayer.ActionLocked(Main.localPlayer, VAction.DA),
+                () => {
                     Main.PlaySound(22, -1, -1, 1);
                     IHPlayer.ToggleActionLock(Main.localPlayer, VAction.DA);
                     Buttons[VAction.DA].Update();
                 },
                 new Vector2(posX, posY) );
 
-            // lockQS = new IHToggle("qslock", "Quick Stack Locked", "Quick Stack Unlocked", null, () => {return IHPlayer.qsLocked;}, () =>
-            // Buttons[VAction.QS] = new IHToggle("QS", "qs", null, () => IHPlayer.ActionLocked(Main.localPlayer, VAction.QS), () => //onToggle
-            Buttons[VAction.QS] = new IHToggle("Quick Stack Locked", "Quick Stack Unlocked", mbase.textures["resources/btn_quickstack"], () => IHPlayer.ActionLocked(Main.localPlayer, VAction.QS), () => //onToggle
-            {
+            Buttons[VAction.QS] = new IHToggle("Quick Stack Locked", "Quick Stack Unlocked",
+                mbase.textures["resources/btn_quickstack"],
+                () => IHPlayer.ActionLocked(Main.localPlayer, VAction.QS),
+                () => {
                     Main.PlaySound(22, -1, -1, 1);
                     IHPlayer.ToggleActionLock(Main.localPlayer, VAction.QS);
                     Buttons[VAction.QS].Update();
                 },
                 new Vector2(posX, (Main.inventoryBackTexture.Height * Main.inventoryScale) + posY) );
+
 
             // lockLA = new IHToggle("lalock", "Loot All Locked", "Loot All Unlocked", null, () => {return IHPlayer.laLocked;}, () =>
             // Buttons[VAction.LA] =  new IHToggle("LA", "la", null, () => IHPlayer.ActionLocked(Main.localPlayer, VAction.LA), () => //onToggle
